requirements.txt
ipython==2.3.0
Django==1.7.1
gnureadline==6.3.3
Markdown==2.5.1
djangorestframework==2.4.4
django-cors-headers==0.13
